Best-card-for: personal "From your wallet" row above the global picks
Adds a small horizontal row above the /best-card-for/<store>/ picks list:

- Signed in: top 3 cards from the user's own wallet ranked for this store,
  using the same scoring as the global list but with floors lowered so a 1%
  Bilt card still surfaces if it's the best they own.
- Signed out: a quiet CTA pointing to /register — "Sign up free and we'll
  show you the best card you already own to use at <Store>."

Implementation:

- Extracted the existing inline ranking machinery from page.tsx into a
  shared lib at lib/storeRanking.ts so the new client component can call
  the same rankCards function. Page is unchanged behaviorally.
- New client component StorePersonalRow handles the auth + wallet fetch,
  joins wallet rows to full Card objects (by db_card_id then card_name),
  ranks them via rankCards({ flatRateFloor: 0, fillFloor: 0, max: 3 }),
  and renders a 3-column horizontal row (1 column on mobile).
- New CSS section .store-personal-row* in landing.css under .store-v2.
