+++ apps/web-next/src/app/best-card-for/[slug]/page.tsx
@@ -1,13 +1,14 @@
 import { Metadata } from "next";
 import { notFound } from "next/navigation";
 import Link from "next/link";
-import { getAllStores, getStore, type Store } from "@/lib/stores";
-import { getAllCards, type Card, type Reward } from "@/lib/api";
-import { getValuation } from "@/lib/valuations";
+import { getAllStores, getStore } from "@/lib/stores";
+import { getAllCards } from "@/lib/api";
+import { rankCards, formatRate, labelForCategory } from "@/lib/storeRanking";
 import { BreadcrumbSchema, FAQSchema } from "@/components/seo/JsonLd";
 import CardImage from "@/components/ui/CardImage";
 import { V2Footer } from "@/components/landing-v2/Chrome";
 import { PencilSquareIcon, ExclamationTriangleIcon, CalendarDaysIcon } from "@heroicons/react/24/outline";
+import StorePersonalRow from "./StorePersonalRow";
 import "../../landing.css";
 
 interface PageProps {
@@ -34,370 +35,9 @@ export async function generateMetadata({ params }: PageProps): Promise<Metadata>
   };
 }
 
-type MatchMode =
-  | 'direct'              // category bonus that always applies (no activation needed)
-  | 'rotating_current'    // quarterly rotation; this category is in `current_categories`
-  | 'rotating_eligible'   // quarterly rotation; in `eligible_categories` but not currently active
-  | 'user_choice'         // user selects N categories from `eligible_categories`
-  | 'top_spend';          // auto: pays headline rate on whichever eligible category is your top spend
+// Ranking machinery (rankCards, findCategoryMatch, channel maps, etc.) lives in
+// `lib/storeRanking.ts` so the per-wallet client component can share it.
 
-type Channel = 'both' | 'online' | 'in_store';
-
-interface RankedPick {
-  card: Card;
-  rate: number;
-  unit: 'percent' | 'points_per_dollar';
-  effectiveRate: number;  // cashback-equivalent % for honest cross-card ranking
-  reason: string;
-  badge?: string;         // short mode-context tag, e.g. "this quarter" / "if selected"
-  source: 'co_brand' | 'also_earns' | 'category' | 'flat_rate';
-  matchMode?: MatchMode;  // only set when source === 'category'
-  channel?: Channel;      // when this pick's bonus applies; omitted for flat_rate
-  note?: string;
-}
-
-// Convert a rate (in either percent or points-per-dollar) into a single
-// cashback-equivalent percentage so we can compare points and cashback
-// cards apples-to-apples. Hilton 7x at 0.5cpp = 3.5%; Citi Double Cash
-// 2% = 2.0%. Without this, the matcher would surface Hilton Aspire (7x)
-// above flat 3% cards even though the cash value is lower.
-function effectiveCashbackRate(rate: number, unit: 'percent' | 'points_per_dollar', cardName: string): number {
-  if (unit === 'percent') return rate;
-  const cpp = getValuation(cardName); // cents per point
-  return rate * cpp;
-}
-
-// Map of which categories are online-only / in-store-only / both. Most
-// merchant-style categories (department_stores, drugstores, wholesale_clubs)
-// are merchant codes that apply at both registers and the brand's website,
-// so the bonus earns either way. `online_shopping` is online by definition.
-const CATEGORY_CHANNEL: Record<string, Channel> = {
-  online_shopping: 'online',
-  amazon: 'online',
-  rakuten: 'online',
-  rakuten_dining: 'online',
-  travel_portal: 'online',
-  hotels_portal: 'online',
-  flights_portal: 'online',
-  car_rentals_portal: 'online',
-  hotels_car_portal: 'online',
-};
-
-function channelForCategory(categoryId: string): Channel {
-  return CATEGORY_CHANNEL[categoryId] || 'both';
-}
-
-interface CardMatch {
-  reward: Reward;
-  matchedCategory: string;  // which of the store's categories triggered the match
-  mode: MatchMode;
-}
-
-const FLAT_RATE_FLOOR = 1.5;
-const MAX_PICKS = 10;
-
-// Infer the rotation/choice mode from data shape so cards with a missing
-// `mode` field still classify correctly. `current_categories` ⇒ rotating;
-// `choices` ⇒ user_choice; otherwise defer to the explicit `mode` field.
-function inferRewardMode(r: Reward): 'quarterly_rotating' | 'user_choice' | 'auto_top_spend' | 'direct' {
-  if (r.mode === 'quarterly_rotating' || r.current_categories || r.current_period) {
-    return 'quarterly_rotating';
-  }
-  if (r.mode === 'auto_top_spend' || r.category === 'top_category') {
-    return 'auto_top_spend';
-  }
-  if (r.mode === 'user_choice' || typeof r.choices === 'number') {
-    return 'user_choice';
-  }
-  return 'direct';
-}
-
-// Find the highest-value applicable reward on a card for any of the store's
-// categories. Considers direct matches, current rotating bonuses, user-choice
-// eligible categories, and historically-eligible rotating categories — each
-// tagged with a `mode` so the caller can render an honest caveat.
-//
-// `includeMerchantSpecific` opts back in to category-bonuses gated by note
-// (Apple Card 3% online, Costco Visa 2% wholesale_clubs). Pass `true` from
-// the co-brand code path because store.co_brand_cards already establishes
-// "this card earns at THIS store" — so we want to use the merchant-gated
-// rate for display rather than falling back to flat 1%.
-function findCategoryMatch(card: Card, categories: string[], includeMerchantSpecific = false): CardMatch | null {
-  if (!card.rewards) return null;
-  // Priority weight per mode: ensures a 5% direct beats a 5% rotating_eligible
-  // even after they're flattened into the same picks list.
-  const modeRank: Record<MatchMode, number> = {
-    direct: 5,
-    rotating_current: 4,
-    user_choice: 3,
-    top_spend: 2,
-    rotating_eligible: 1,
-  };
-  let best: CardMatch | null = null;
-
-  for (const r of card.rewards) {
-    const inferred = inferRewardMode(r);
-
-    // Direct: the reward's own `category` is one the store maps to AND it's
-    // not actually a rotating/user-choice slot in disguise. Skip when
-    // `merchant_specific` is set — those bonuses are gated by free-text
-    // in `note` (e.g. Apple Card's "select Apple Pay merchants") and
-    // would falsely surface for arbitrary stores in the same category.
-    if (inferred === 'direct' && categories.includes(r.category) && (!r.merchant_specific || includeMerchantSpecific)) {
-      const candidate: CardMatch = { reward: r, matchedCategory: r.category, mode: 'direct' };
-      if (!best || compareMatches(candidate, best, modeRank) > 0) best = candidate;
-      continue;
-    }
-
-    if (inferred === 'quarterly_rotating') {
-      const current = (r.current_categories || []).map(c => typeof c === 'string' ? c : c.category);
-      const eligible = r.eligible_categories || [];
-      const inCurrent = categories.find(c => current.includes(c));
-      const inEligible = !inCurrent ? categories.find(c => eligible.includes(c)) : undefined;
-      if (inCurrent) {
-        const candidate: CardMatch = { reward: r, matchedCategory: inCurrent, mode: 'rotating_current' };
-        if (!best || compareMatches(candidate, best, modeRank) > 0) best = candidate;
-      } else if (inEligible) {
-        const candidate: CardMatch = { reward: r, matchedCategory: inEligible, mode: 'rotating_eligible' };
-        if (!best || compareMatches(candidate, best, modeRank) > 0) best = candidate;
-      }
-      continue;
-    }
-
-    if (inferred === 'auto_top_spend') {
-      const eligible = r.eligible_categories || [];
-      const matched = categories.find(c => eligible.includes(c));
-      if (matched) {
-        const candidate: CardMatch = { reward: r, matchedCategory: matched, mode: 'top_spend' };
-        if (!best || compareMatches(candidate, best, modeRank) > 0) best = candidate;
-      }
-      continue;
-    }
-
-    if (inferred === 'user_choice') {
-      const eligible = r.eligible_categories || [];
-      const matched = categories.find(c => eligible.includes(c));
-      if (matched) {
-        const candidate: CardMatch = { reward: r, matchedCategory: matched, mode: 'user_choice' };
-        if (!best || compareMatches(candidate, best, modeRank) > 0) best = candidate;
-      }
-    }
-  }
-
-  return best;
-}
-
-function compareMatches(a: CardMatch, b: CardMatch, modeRank: Record<MatchMode, number>): number {
-  if (a.reward.value !== b.reward.value) return a.reward.value - b.reward.value;
-  return modeRank[a.mode] - modeRank[b.mode];
-}
-
-function flatRateReward(card: Card): Reward | null {
-  if (!card.rewards) return null;
-  return card.rewards.find(r => r.category === 'everything_else') || null;
-}
-
-function formatRate(value: number, unit: 'percent' | 'points_per_dollar'): string {
-  if (unit === 'percent') return `${value}%`;
-  return `${value}x points`;
-}
-
-function formatCap(reward: Reward): string {
-  if (!reward.spend_cap || !reward.cap_period) return '';
-  const period = reward.cap_period === 'quarterly' ? 'quarter'
-    : reward.cap_period === 'monthly' ? 'month'
-    : reward.cap_period === 'annual' ? 'year'
-    : reward.cap_period;
-  const after = reward.rate_after_cap !== undefined ? `, then ${reward.rate_after_cap}%` : '';
-  return ` (up to $${reward.spend_cap.toLocaleString()}/${period}${after})`;
-}
-
-function reasonAndBadgeForMatch(match: CardMatch): { reason: string; badge: string } {
-  const rateStr = formatRate(match.reward.value, match.reward.unit as 'percent' | 'points_per_dollar');
-  const catLabel = labelForCategory(match.matchedCategory);
-  const cap = formatCap(match.reward);
-  switch (match.mode) {
-    case 'direct':
-      return { reason: `${rateStr} on ${catLabel}${cap}`, badge: '' };
-    case 'rotating_current': {
-      const period = match.reward.current_period ? ` (${match.reward.current_period})` : '';
-      return {
-        reason: `${rateStr} on ${catLabel} this quarter${period}${cap}. Activation required each quarter.`,
-        badge: 'this quarter',
-      };
-    }
-    case 'user_choice':
-      return {
-        reason: `${rateStr} on ${catLabel} if you select it as a bonus category${cap}`,
-        badge: 'if you select it',
-      };
-    case 'top_spend':
-      return {
-        reason: `${rateStr} on ${catLabel} if it's your top eligible spend category that cycle${cap}`,
-        badge: 'if it’s your top category',
-      };
-    case 'rotating_eligible':
-      return {
-        reason: `Up to ${rateStr} on ${catLabel} when it rotates in. Not in this quarter's lineup — check before a trip.`,
-        badge: 'situational',
-      };
-  }
-}
-
-function rankCards(store: Store, cards: Card[]): RankedPick[] {
-  const active = cards.filter(c => c.accepting_applications !== false);
-  const cardsBySlug = new Map(active.map(c => [c.slug, c]));
-  const used = new Set<string>();
-  const picks: RankedPick[] = [];
-
-  // 1. Co-brand
-  for (const slug of store.co_brand_cards || []) {
-    const card = cardsBySlug.get(slug);
-    if (!card || used.has(slug)) continue;
-    // Opt into merchant-gated rewards on the co-brand path. Costco Visa's
-    // 2% wholesale_clubs is `merchant_specific` (so it doesn't leak onto
-    // BJ's / Sam's), but for the Costco store page we WANT to show that
-    // 2% rate rather than fall back to flat 1%.
-    const match = findCategoryMatch(card, store.categories, true);
-    const r = match?.reward || flatRateReward(card);
-    const rate = r?.value ?? 0;
-    const unit = (r?.unit as 'percent' | 'points_per_dollar') ?? 'percent';
-    picks.push({
-      card,
-      rate,
-      unit,
-      effectiveRate: effectiveCashbackRate(rate, unit, card.card_name),
-      reason: `Co-branded ${store.name} card`,
-      source: 'co_brand',
-      channel: 'both',  // co-brand cards earn at the brand both in-store and online
-      note: r?.note,
-    });
-    used.add(slug);
-  }
-
-  // 2. Build the rate-ranked group: merchant-specific overrides (also_earns)
-  //    + category bonuses (direct / rotating_current / user_choice /
-  //    rotating_eligible) compete in the same list, sorted by effective
-  //    rate. Previously also_earns lived in its own tier above category
-  //    matches, which meant a 5% also_earns would outrank a 6% direct
-  //    grocery match (e.g. Amazon Prime Visa at Whole Foods vs. Blue Cash
-  //    Preferred). Sorting them together is the honest answer at-a-glance.
-  type RankedCandidate =
-    | { kind: 'also_earns'; card: Card; rate: number; unit: 'percent' | 'points_per_dollar'; note?: string; effective: number }
-    | { kind: 'category'; card: Card; match: CardMatch; effective: number };
-  const candidates: RankedCandidate[] = [];
-
-  for (const entry of store.also_earns || []) {
-    const card = cardsBySlug.get(entry.card);
-    if (!card || used.has(entry.card)) continue;
-    const eff = effectiveCashbackRate(entry.rate, entry.unit, card.card_name);
-    candidates.push({
-      kind: 'also_earns',
-      card,
-      rate: entry.rate,
-      unit: entry.unit,
-      note: entry.note,
-      effective: eff,
-    });
-  }
-
-  for (const card of active) {
-    if (used.has(card.slug)) continue;
-    if (candidates.some(c => c.card.slug === card.slug)) continue; // already an also_earns pick
-    const m = findCategoryMatch(card, store.categories);
-    if (!m) continue;
-    const eff = effectiveCashbackRate(
-      m.reward.value,
-      m.reward.unit as 'percent' | 'points_per_dollar',
-      card.card_name,
-    );
-    // Skip anything below the flat-rate floor on a cashback-equivalent
-    // basis. Hilton 1x at "everything else" = 0.5% effective — below floor.
-    if (eff <= FLAT_RATE_FLOOR) continue;
-    // rotating_eligible is "you might earn this in some quarter" — penalize
-    // so it lists under everything that's actually current/choosable.
-    const effective = m.mode === 'rotating_eligible' ? eff - 100 : eff;
-    candidates.push({ kind: 'category', card, match: m, effective });
-  }
-
-  candidates.sort((a, b) => b.effective - a.effective);
-
-  for (const c of candidates) {
-    if (c.kind === 'also_earns') {
-      picks.push({
-        card: c.card,
-        rate: c.rate,
-        unit: c.unit,
-        effectiveRate: effectiveCashbackRate(c.rate, c.unit, c.card.card_name),
-        reason: `Earns ${formatRate(c.rate, c.unit)} at ${store.name}`,
-        source: 'also_earns',
-        channel: 'both',
-        note: c.note,
-      });
-    } else {
-      const { reason, badge } = reasonAndBadgeForMatch(c.match);
-      picks.push({
-        card: c.card,
-        rate: c.match.reward.value,
-        unit: c.match.reward.unit as 'percent' | 'points_per_dollar',
-        effectiveRate: effectiveCashbackRate(
-          c.match.reward.value,
-          c.match.reward.unit as 'percent' | 'points_per_dollar',
-          c.card.card_name,
-        ),
-        reason,
-        badge: badge || undefined,
-        channel: channelForCategory(c.match.matchedCategory),
-        source: 'category',
-        matchMode: c.match.mode,
-        note: c.match.reward.note,
-      });
-    }
-    used.add(c.card.slug);
-  }
-
-  // 4. Flat-rate fallback to fill the list out to MAX_PICKS. Always runs
-  //    so a store with only 1-2 category matches still shows a useful
-  //    top-10. Floor at 2% so we don't list every 1% card on the planet.
-  if (picks.length < MAX_PICKS) {
-    const flatPicks: { card: Card; reward: Reward }[] = [];
-    for (const card of active) {
-      if (used.has(card.slug)) continue;
-      const reward = flatRateReward(card);
-      if (reward && reward.unit === 'percent' && reward.value >= 2) {
-        flatPicks.push({ card, reward });
-      }
-    }
-    flatPicks.sort((a, b) => b.reward.value - a.reward.value);
-    for (const { card, reward } of flatPicks.slice(0, MAX_PICKS - picks.length)) {
-      picks.push({
-        card,
-        rate: reward.value,
-        unit: reward.unit as 'percent',
-        effectiveRate: reward.value, // already a cashback %
-        reason: `${formatRate(reward.value, 'percent')} flat-rate cashback`,
-        source: 'flat_rate',
-        note: reward.note,
-      });
-      used.add(card.slug);
-    }
-  }
-
-  return picks.slice(0, MAX_PICKS);
-}
-
-const CATEGORY_LABELS: Record<string, string> = {
-  department_stores: 'department stores',
-  online_shopping: 'online shopping',
-  groceries: 'groceries',
-  dining: 'dining',
-  gas: 'gas',
-  travel: 'travel',
-  everything_else: 'everything else',
-  home_improvement: 'home improvement',
-  drugstores: 'drugstores',
-  wholesale_clubs: 'wholesale clubs',
-};
 
 // Tag form: shorter, title-cased, singular where natural
 // (e.g. "Department Store" instead of "department stores").
@@ -412,10 +52,6 @@ const CATEGORY_TAG_LABELS: Record<string, string> = {
   drugstores: 'Drugstore',
   wholesale_clubs: 'Wholesale Club',
 };
-
-function labelForCategory(id: string): string {
-  return CATEGORY_LABELS[id] || id.replace(/_/g, ' ');
-}
 
 function tagLabelForCategory(id: string): string {
   return CATEGORY_TAG_LABELS[id]
@@ -477,6 +113,7 @@ export default async function BestCardForStorePage({ params }: PageProps) {
           </p>
         ) : (
           <>
+          <StorePersonalRow store={store} cards={allCards} />
           {picks.some(p => p.channel === 'online' || p.channel === 'in_store') && (
             <p className="store-channel-note">
               Picks without a tag earn at {store.name} both in-store and online. Watch for the
